Map can be downloaded with the PDF report and printed (#99)
It is still a basic representation of the map. Google only allows free users to generate 640x640 pixel images. If we were paid users (a couple of thousand dollars a year) we would be able to generate 2048x2048 pixel images.

Next step is to try to generate more complete and detailed map images from the open street map api.